requirements.txt
openai>=1.63.0
python-dotenv>=1.0.0
requests>=2.31.0
weaviate-client==4.17.0
streamlit>=1.30.0